+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Prepared for Grafana 11 (#399)
 - Updated Reset button handler (#422)
 - Updated Confirmation Window (#420)
+- Updated Error handling in code editors showIf, disableIf, getOptions (#410)
 
 ## 3.8.0 (2024-05-30)